backend/tests/customer_portal/test_authentication.py
"""
Tests for customer authentication and login serializer.
"""

import pytest
from rest_framework import status
from rest_framework.test import APIClient

from apps.accounts.models import Company, CustomerProfile, CustomUser
from apps.accounts.serializers import UnifiedLoginSerializer


@pytest.fixture
def company(db):
    """Create test company."""
    return Company.objects.create(name="Test Company", slug="test-company")


@pytest.fixture
def customer_user(db, company):
    """Create a customer user with username and password."""
    user = CustomUser.objects.create_user(
        username="testcustomer",
        password="password123",
        user_type="customer",
        first_name="Test",
        last_name="Customer",
    )
    CustomerProfile.objects.create(
        user=user, company=company, phone_number="+998901234567"
    )
    return user


@pytest.fixture
def manager_user(db):
    """Create a manager user."""
    return CustomUser.objects.create_user(
        username="testmanager",
        password="password123",
        user_type="manager",
        phone_number="+998907654321",
    )


@pytest.fixture
def admin_user(db):
    """Create an admin user."""
    return CustomUser.objects.create_user(
        username="testadmin",
        password="password123",
        user_type="admin",
        is_staff=True,
    )


@pytest.fixture
def api_client():
    """Create API client."""
    return APIClient()


@pytest.mark.django_db
class TestUnifiedLoginSerializer:
    """Tests for UnifiedLoginSerializer with customer support."""

    def test_detects_customer_user_type(self, customer_user):
        """Test that serializer correctly identifies customer user type."""
        serializer = UnifiedLoginSerializer(
            data={"login": "testcustomer", "password": "password123"}
        )

        assert serializer.is_valid() is True
        assert serializer.validated_data["user"] == customer_user
        assert serializer.validated_data["user_type"] == "customer"

    def test_detects_manager_user_type_by_phone(self, manager_user):
        """Test that serializer identifies manager by phone number."""
        serializer = UnifiedLoginSerializer(
            data={"login": "+998907654321", "password": "password123"}
        )

        assert serializer.is_valid() is True
        assert serializer.validated_data["user"] == manager_user
        assert serializer.validated_data["user_type"] == "manager"

    def test_detects_admin_user_type_by_username(self, admin_user):
        """Test that serializer identifies admin by username."""
        serializer = UnifiedLoginSerializer(
            data={"login": "testadmin", "password": "password123"}
        )

        assert serializer.is_valid() is True
        assert serializer.validated_data["user"] == admin_user
        assert serializer.validated_data["user_type"] == "custom_user"

    def test_rejects_wrong_password(self, customer_user):
        """Test that serializer rejects wrong password."""
        from rest_framework.exceptions import AuthenticationFailed

        serializer = UnifiedLoginSerializer(
            data={"login": "testcustomer", "password": "wrongpassword"}
        )

        # Serializer raises AuthenticationFailed instead of returning False
        with pytest.raises(AuthenticationFailed):
            serializer.is_valid(raise_exception=True)

    def test_rejects_nonexistent_user(self):
        """Test that serializer rejects nonexistent user."""
        from rest_framework.exceptions import AuthenticationFailed

        serializer = UnifiedLoginSerializer(
            data={"login": "nonexistent", "password": "password123"}
        )

        # Serializer raises AuthenticationFailed instead of returning False
        with pytest.raises(AuthenticationFailed):
            serializer.is_valid(raise_exception=True)


@pytest.mark.django_db
class TestCustomerLoginAPI:
    """Tests for customer login API endpoint."""

    def test_customer_login_returns_correct_response(self, api_client, customer_user):
        """Test that customer login returns customer-specific response."""
        response = api_client.post(
            "/api/auth/login/",
            {"login": "testcustomer", "password": "password123"},
            format="json",
        )

        assert response.status_code == status.HTTP_200_OK
        assert response.data["success"] is True
        assert response.data["user_type"] == "customer"

        # Check JWT tokens
        assert "access" in response.data
        assert "refresh" in response.data

        # Check user data
        user_data = response.data["user"]
        assert user_data["id"] == customer_user.id
        assert user_data["username"] == "testcustomer"
        assert user_data["full_name"] == "Test Customer"
        assert user_data["phone_number"] == "+998901234567"

        # Check company data
        assert "company" in user_data
        assert user_data["company"]["name"] == "Test Company"
        assert user_data["company"]["slug"] == "test-company"

    def test_manager_login_returns_manager_response(self, api_client, manager_user):
        """Test that manager login returns manager-specific response."""
        response = api_client.post(
            "/api/auth/login/",
            {"login": "+998907654321", "password": "password123"},
            format="json",
        )

        assert response.status_code == status.HTTP_200_OK
        assert response.data["user_type"] == "manager"

    def test_admin_login_returns_admin_response(self, api_client, admin_user):
        """Test that admin login returns admin-specific response."""
        response = api_client.post(
            "/api/auth/login/",
            {"login": "testadmin", "password": "password123"},
            format="json",
        )

        assert response.status_code == status.HTTP_200_OK
        assert response.data["user_type"] == "custom_user"

    def test_jwt_token_contains_user_type(self, api_client, customer_user):
        """Test that JWT token contains user_type claim."""
        response = api_client.post(
            "/api/auth/login/",
            {"login": "testcustomer", "password": "password123"},
            format="json",
        )

        # Decode JWT to verify user_type claim
        import jwt
        from django.conf import settings

        token = response.data["access"]
        decoded = jwt.decode(
            token, settings.SECRET_KEY, algorithms=["HS256"], options={"verify_exp": False}
        )

        assert decoded["user_type"] == "customer"
        assert decoded["user_id"] == str(customer_user.id)

    def test_customer_can_use_token_to_access_endpoints(
        self, api_client, customer_user
    ):
        """Test that customer can use JWT token to access protected endpoints."""
        # Login to get token
        login_response = api_client.post(
            "/api/auth/login/",
            {"login": "testcustomer", "password": "password123"},
            format="json",
        )

        token = login_response.data["access"]

        # Use token to access customer endpoint
        api_client.credentials(HTTP_AUTHORIZATION=f"Bearer {token}")
        profile_response = api_client.get("/api/customer/profile/")

        assert profile_response.status_code == status.HTTP_200_OK
        assert profile_response.data["success"] is True

    def test_inactive_customer_cannot_login(self, api_client, customer_user):
        """Test that inactive customer cannot login."""
        customer_user.is_active = False
        customer_user.save()

        response = api_client.post(
            "/api/auth/login/",
            {"login": "testcustomer", "password": "password123"},
            format="json",
        )

        assert response.status_code == status.HTTP_403_FORBIDDEN

    def test_logout_blacklists_token(self, api_client, customer_user):
        """Test that logout blacklists the refresh token."""
        # Login
        login_response = api_client.post(
            "/api/auth/login/",
            {"login": "testcustomer", "password": "password123"},
            format="json",
        )

        token = login_response.data["access"]
        refresh_token = login_response.data["refresh"]

        # Logout
        api_client.credentials(HTTP_AUTHORIZATION=f"Bearer {token}")
        logout_response = api_client.post(
            "/api/auth/logout/", {"refresh": refresh_token}, format="json"
        )

        assert logout_response.status_code == status.HTTP_200_OK


@pytest.mark.django_db
class TestCustomerLoginEdgeCases:
    """Tests for edge cases in customer authentication."""

    def test_customer_without_profile_can_login(self, db):
        """Test that customer without CustomerProfile can still login (legacy support)."""
        # Create customer without profile
        user = CustomUser.objects.create_user(
            username="legacycustomer",
            password="password123",
            user_type="customer",
            first_name="Legacy",
            last_name="Customer",
        )

        api_client = APIClient()
        response = api_client.post(
            "/api/auth/login/",
            {"login": "legacycustomer", "password": "password123"},
            format="json",
        )

        assert response.status_code == status.HTTP_200_OK
        assert response.data["user_type"] == "customer"
        # Company should be None for legacy users
        assert response.data["user"]["company"] is None

    def test_customer_with_username_starting_with_digit(self, db, company):
        """Test customer with username that looks like phone number."""
        user = CustomUser.objects.create_user(
            username="123customer",
            password="password123",
            user_type="customer",
        )
        CustomerProfile.objects.create(
            user=user, company=company, phone_number="+998901111111"
        )

        # Should try phone auth first, fail, then username auth
        api_client = APIClient()
        response = api_client.post(
            "/api/auth/login/",
            {"login": "123customer", "password": "password123"},
            format="json",
        )

        # This should fail because it tries phone auth first
        # This is expected behavior - usernames should not start with digits
        # Custom exception handler returns 403 for authentication failures
        assert response.status_code in [status.HTTP_401_UNAUTHORIZED, status.HTTP_403_FORBIDDEN]
